+++ DateTime.d.ts
@@ -35,12 +35,17 @@ declare namespace ReactDatetimeClass {
          Represents the selected date by the component, in order to use it as a controlled component.
          This prop is parsed by moment.js, so it is possible to use a date string or a moment.js date.
          */
-        value?: Date;
+        value?: Date | string | Moment;
         /*
          Represents the selected date for the component to use it as a uncontrolled component.
          This prop is parsed by moment.js, so it is possible to use a date string or a moment.js date.
          */
-        defaultValue?: Date;
+        defaultValue?: Date | string | Moment;
+        /*
+         Represents the month which is viewed on opening the calendar when there is no selected date.
+         This prop is parsed by Moment.js, so it is possible to use a date `string` or a `moment` object.
+         */
+        viewDate?: Date | string | Moment;
         /*
          Represents the default date for the picker when the initial value is not set.
          This prop is parsed by moment.js, so it is possible to use a date string or a moment.js date.
@@ -51,13 +56,13 @@ declare namespace ReactDatetimeClass {
          If true the date will be displayed using the defaults for the current locale.
          If false the datepicker is disabled and the component can be used as timepicker.
          */
-        dateFormat?: boolean|string;
+        dateFormat?: boolean | string;
         /*
          Defines the format for the time. It accepts any moment.js time format.
          If true the time will be displayed using the defaults for the current locale.
          If false the timepicker is disabled and the component can be used as datepicker.
          */
-        timeFormat?: boolean|string;
+        timeFormat?: boolean | string;
         /*
          Whether to show an input field to edit the date manually.
          */
@@ -77,11 +82,16 @@ declare namespace ReactDatetimeClass {
          */
         utc?: boolean;
         /*
+         When specified, input time values will be displayed in the given time zone. Otherwise they will default
+         to the user's local timezone (unless `utc` specified).
+         */
+        displayTimeZone?: string;
+        /*
          Callback trigger when the date changes. The callback receives the selected `moment` object as
          only parameter, if the date in the input is valid. If the date in the input is not valid, the
          callback receives the value of the input (a string).
          */
-        onChange?: EventOrValueHandler<ChangeEvent<any>>;
+        onChange?: (value: Moment | string) => void;
         /*
          Callback trigger for when the user opens the datepicker.
          */
@@ -99,9 +109,19 @@ declare namespace ReactDatetimeClass {
          */
         onViewModeChange?: (viewMode: string) => void;
         /*
+         Callback trigger when the user navigates to the previous month, year or decade.
+         The callback receives the amount and type ('month', 'year') as parameters.
+         */
+        onNavigateBack?: (amount: number, type: string) => void;
+        /*
+         Callback trigger when the user navigates to the next month, year or decade.
+         The callback receives the amount and type ('month', 'year') as parameters.
+         */
+        onNavigateForward?: (amount: number, type: string) => void;
+        /*
          The default view to display when the picker is shown. ('years', 'months', 'days', 'time')
          */
-        viewMode?: ViewMode|number;
+        viewMode?: ViewMode | number;
         /*
          Extra class names for the component markup.
          */